backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.46.3
torch==2.5.1
PyMuPDF==1.24.11
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0
sentencepiece==0.2.0
accelerate==1.1.1
